remove redundant build function and activate workspace on build failure.
